Create Mechanoreceptors.cpp
In the expanded version, I've included the following:

Member variables: You can add specific variables to represent the state or properties of the mechanoreceptors. These variables would hold information or data relevant to detecting vibrations and sensing air currents.

Constructors: The Mechanoreceptors class includes a default constructor, which can be customized to initialize the member variables or perform any necessary setup when creating an instance of the mechanoreceptors.

Destructor: The destructor is responsible for any necessary cleanup or deallocation of resources associated with the Mechanoreceptors class.

Member functions: I've added two member functions to illustrate some common functions of the mechanoreceptors: detectVibrations() and senseAirCurrents(). You can further expand these functions to implement the specific mechanisms for detecting vibrations and sensing air currents using the mechanoreceptors of the jumping spider.

Additional member functions: You can add any additional member functions as needed to support the functionality of the Mechanoreceptors class. These functions could handle more specific mechanosensory tasks or behaviors associated with the spider's mechanoreceptors.

Remember, this is a simplified representation, and you can extend and customize the Mechanoreceptors class based on the specific requirements and behaviors of the jumping spider you're modeling.