Create ChemoThermoreceptors.cpp
In the expanded version, I've included the following:

Member variables: You can add specific variables to represent the state or properties of the chemo- and thermoreceptors. These variables would hold information or data relevant to detecting chemical signals and sensing temperature variations.

Constructors: The ChemoThermoreceptors class includes a default constructor, which can be customized to initialize the member variables or perform any necessary setup when creating an instance of the chemo- and thermoreceptors.

Destructor: The destructor is responsible for any necessary cleanup or deallocation of resources associated with the ChemoThermoreceptors class.

Member functions: I've added two member functions to illustrate some common functions of the chemo- and thermoreceptors: detectChemicalSignals() and senseTemperatureVariations(). You can further expand these functions to implement the specific mechanisms for detecting chemical signals and sensing temperature variations using the chemo- and thermoreceptors of the jumping spider.

Additional member functions: You can add any additional member functions as needed to support the functionality of the ChemoThermoreceptors class. These functions could handle more specific chemosensory or thermosensory tasks or behaviors associated with the spider's chemo- and thermoreceptors.

Remember, this is a simplified representation, and you can extend and customize the ChemoThermoreceptors class based on the specific requirements and behaviors of the jumping spider you're modeling.