Create pram_algorithms.cpp
Sure! Here's the cleaned copypasta and a C++ implementation:

Copypasta:
A parallel random access machine (PRAM) consists of a number of processors that work in parallel and use a shared memory. The memory is comprised of memory cells which can be accessed by a processor in O(1) time. Furthermore, we assume that simple arithmetic and bitwise operations, including addition, can be done in O(1) time by a processor. We mostly use the Concurrent-Read ConcurrentWrite model (CRCW PRAM), i.e. processors are allowed to read and write concurrently from and to the same memory location. More precisely, we assume the common PRAM model: several processors can concurrently write into the same memory location, only if all of them write the same value. We also mention the Exclusive-Read Exclusive-Write model (EREW PRAM), where concurrent access is not allowed. The work of a PRAM computation is the sum of the number of all computation steps of all processors made during the computation. We define the space s required by a PRAM computation as the maximal index of any memory cell accessed during the computation. We refer to [9] for more details on PRAMs and to [13, Section 2.2.3] for a discussion of alternative space measures. The main feature of the common CRCW model relevant for our algorithms that separates it from the EREW model is that it allows to compute the minimum or maximum value of an array of size n in constant time (with work O(n1+ϵ)) which is shown in another paper at MFCS 2023.3. For simplicity, we assume that even if the size bound n grows, a number in the range [0, n] can still be stored in one memory cell. This assumption is justified, since addition of larger numbers N can still be done in constant time and polylogarithmic work on a CRCW PRAM. Additionally, we assume that the number of processors always depends on the current size bound n. Hence, the number of processors increases with growing n which allows us to use the PRAM model with growing structures. We describe our PRAM algorithms on an abstract level and do not exactly specify how processors are assigned to data. Whenever an algorithm does something in parallel for a set of objects, these objects can be assigned to a bunch of processors with the help of some underlying array. This is relatively straightforward for strings and substrings and the data structures used in Section 4. In Section 3, it is usually based on zone records and their underlying partition records.

C++ Implementation:

Please note that the PRAM computation function (PRAMComputation) should be replaced with the actual implementation of your PRAM algorithm. This example shows a basic framework for running a PRAM computation with a given size bound n and the number of processors processors. You can replace the implementation inside PRAMComputation with your specific PRAM algorithm logic.